Create gatherStarStats.py
User provides a variable star using a known ID of the object. The program prints the equatorial coordinates in decimal degrees, a common identifier of the object in two formats, and the magnitude of the object. The program then gathers all of the stars within a user-provided radius, and picks the one closest in magnitude to the variable star to be a reference star. The program finally outputs the same information for the reference star as it did for the variable star.